Move respective deserialization methods from BridgeSerializationUtils to PendingFederation and FederationMember. Move respective tests from BridgeSerializationUtilsTest to PendingFederationTest and FederationMemberTest. Rename methods to be clearer.
