+++ rskj-core/src/main/java/co/rsk/peg/federation/FederationMember.java
@@ -22,6 +22,7 @@ import co.rsk.bitcoinj.core.BtcECKey;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.util.ByteUtil;
 import org.ethereum.util.RLP;
+import org.ethereum.util.RLPList;
 
 import java.util.Arrays;
 import java.util.List;
@@ -38,10 +39,11 @@ import java.util.stream.Collectors;
  */
 public final class FederationMember {
     public static final FederationMemberPubKeysComparator BTC_RSK_MST_PUBKEYS_COMPARATOR = new FederationMemberPubKeysComparator();
-    private static final int FEDERATION_MEMBER_LIST_SIZE = 3;
-    private static final int FEDERATION_MEMBER_BTC_KEY_INDEX = 0;
-    private static final int FEDERATION_MEMBER_RSK_KEY_INDEX = 1;
-    private static final int FEDERATION_MEMBER_MST_KEY_INDEX = 2;
+    private static final int FEDERATION_RLP_LIST_SIZE = 3;
+    private static final int LIST_SIZE = 3;
+    private static final int BTC_KEY_INDEX = 0;
+    private static final int RSK_KEY_INDEX = 1;
+    private static final int MST_KEY_INDEX = 2;
     private final BtcECKey btcPublicKey;
     private final ECKey rskPublicKey;
     private final ECKey mstPublicKey;
@@ -161,12 +163,28 @@ public final class FederationMember {
     }
 
     public byte[] serialize() {
-        byte[][] rlpElements = new byte[FEDERATION_MEMBER_LIST_SIZE][];
-        rlpElements[FEDERATION_MEMBER_BTC_KEY_INDEX] = RLP.encodeElement(
-            this.getBtcPublicKey().getPubKeyPoint().getEncoded(true)
-        );
-        rlpElements[FEDERATION_MEMBER_RSK_KEY_INDEX] = RLP.encodeElement(this.getRskPublicKey().getPubKey(true));
-        rlpElements[FEDERATION_MEMBER_MST_KEY_INDEX] = RLP.encodeElement(this.getMstPublicKey().getPubKey(true));
+        byte[][] rlpElements = new byte[LIST_SIZE][];
+        byte[] btcPublicKeyBytes = this.getBtcPublicKey().getPubKeyPoint().getEncoded(true);
+        byte[] rskPublicKeyBytes = this.getRskPublicKey().getPubKey(true);
+        byte[] mstPublicKeyBytes = this.getMstPublicKey().getPubKey(true);
+
+        rlpElements[BTC_KEY_INDEX] = RLP.encodeElement(btcPublicKeyBytes);
+        rlpElements[RSK_KEY_INDEX] = RLP.encodeElement(rskPublicKeyBytes);
+        rlpElements[MST_KEY_INDEX] = RLP.encodeElement(mstPublicKeyBytes);
         return RLP.encodeList(rlpElements);
+    }
+
+    public static FederationMember deserialize(byte[] data) {
+        RLPList rlpList = (RLPList)RLP.decode2(data).get(0);
+
+        if (rlpList.size() != FEDERATION_RLP_LIST_SIZE) {
+            throw new RuntimeException(String.format("Invalid serialized FederationMember. Expected %d elements but got %d", LIST_SIZE, rlpList.size()));
+        }
+
+        BtcECKey btcKey = BtcECKey.fromPublicOnly(rlpList.get(BTC_KEY_INDEX).getRLPData());
+        ECKey rskKey = ECKey.fromPublicOnly(rlpList.get(RSK_KEY_INDEX).getRLPData());
+        ECKey mstKey = ECKey.fromPublicOnly(rlpList.get(MST_KEY_INDEX).getRLPData());
+
+        return new FederationMember(btcKey, rskKey, mstKey);
     }
 }
