+++ rskj-core/src/test/java/co/rsk/peg/federation/PendingFederationTest.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package co.rsk.peg;
+package co.rsk.peg.federation;
 
 import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.core.NetworkParameters;
@@ -24,12 +24,15 @@ import co.rsk.config.BridgeConstants;
 import co.rsk.config.BridgeMainNetConstants;
 import co.rsk.config.BridgeTestNetConstants;
 import co.rsk.crypto.Keccak256;
-import co.rsk.peg.federation.*;
 import co.rsk.peg.resources.TestConstants;
+import org.ethereum.TestUtils;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.config.blockchain.upgrades.ConsensusRule;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.crypto.HashUtil;
+import org.ethereum.util.ByteUtil;
+import org.ethereum.util.RLP;
+import org.ethereum.util.RLPList;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -40,6 +43,8 @@ import org.mockito.quality.Strictness;
 
 import java.math.BigInteger;
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -247,6 +252,137 @@ class PendingFederationTest {
         assertEquals(expectedHash, pendingFederation.getHash());
     }
 
+    @Test
+    void serializeAndDeserializePendingFederation() {
+        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
+        // we want serialization from members
+        when(activations.isActive(ConsensusRule.RSKIP123)).thenReturn(true);
+
+        final int NUM_CASES = 20;
+
+        for (int i = 0; i < NUM_CASES; i++) {
+            int numMembers = randomInRange(2, 14);
+            List<FederationMember> members = new ArrayList<>();
+            for (int j = 0; j < numMembers; j++) {
+                members.add(new FederationMember(new BtcECKey(), new ECKey(), new ECKey()));
+            }
+            PendingFederation testPendingFederation = new PendingFederation(members);
+
+            byte[] serializedTestPendingFederation = testPendingFederation.serialize(activations);
+            PendingFederation deserializedTestPendingFederation = PendingFederation.deserialize(serializedTestPendingFederation);
+
+            Assertions.assertEquals(testPendingFederation, deserializedTestPendingFederation);
+        }
+    }
+
+    @Test
+    void serializePendingFederation_serializedKeysAreCompressedAndThree() {
+        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
+        // we want serialization from members
+        when(activations.isActive(ConsensusRule.RSKIP123)).thenReturn(true);
+
+        final int NUM_MEMBERS = 10;
+        final int EXPECTED_NUM_KEYS = 3;
+        final int EXPECTED_PUBLICKEY_SIZE = 33;
+
+        List<FederationMember> members = new ArrayList<>();
+        for (int j = 0; j < NUM_MEMBERS; j++) {
+            members.add(new FederationMember(new BtcECKey(), new ECKey(), new ECKey()));
+        }
+
+        PendingFederation testPendingFederation = new PendingFederation(members);
+        byte[] serializedPendingFederation = testPendingFederation.serialize(activations);
+
+        RLPList memberList = (RLPList) RLP.decode2(serializedPendingFederation).get(0);
+        Assertions.assertEquals(NUM_MEMBERS, memberList.size());
+
+        for (int i = 0; i < NUM_MEMBERS; i++) {
+            RLPList memberKeys = (RLPList) RLP.decode2(memberList.get(i).getRLPData()).get(0);
+            Assertions.assertEquals(EXPECTED_NUM_KEYS, memberKeys.size());
+            for (int j = 0; j < EXPECTED_NUM_KEYS; j++) {
+                Assertions.assertEquals(EXPECTED_PUBLICKEY_SIZE, memberKeys.get(j).getRLPData().length);
+            }
+
+        }
+    }
+
+    @Test
+    void deserializePendingFederation_invalidFederationMember() {
+        byte[] serialized = RLP.encodeList(
+            RLP.encodeList(RLP.encodeElement(new byte[0]), RLP.encodeElement(new byte[0]))
+        );
+
+        try {
+            PendingFederation.deserialize(serialized);
+            Assertions.fail();
+        } catch (RuntimeException e) {
+            Assertions.assertTrue(e.getMessage().contains("Invalid serialized FederationMember"));
+        }
+    }
+
+    @Test
+    void serializePendingFederationOnlyBtcKeys() throws Exception {
+        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
+        // we want serialization from pub keys
+        when(activations.isActive(ConsensusRule.RSKIP123)).thenReturn(false);
+
+        byte[][] publicKeyBytes = new byte[][]{
+            BtcECKey.fromPrivate(BigInteger.valueOf(100)).getPubKey(),
+            BtcECKey.fromPrivate(BigInteger.valueOf(200)).getPubKey(),
+            BtcECKey.fromPrivate(BigInteger.valueOf(300)).getPubKey(),
+            BtcECKey.fromPrivate(BigInteger.valueOf(400)).getPubKey(),
+            BtcECKey.fromPrivate(BigInteger.valueOf(500)).getPubKey(),
+            BtcECKey.fromPrivate(BigInteger.valueOf(600)).getPubKey(),
+        };
+
+        // Only actual keys serialized are BTC keys, so we don't really care about RSK or MST keys
+        List<BtcECKey> keys = Arrays.asList(new BtcECKey[]{
+            BtcECKey.fromPublicOnly(publicKeyBytes[0]),
+            BtcECKey.fromPublicOnly(publicKeyBytes[1]),
+            BtcECKey.fromPublicOnly(publicKeyBytes[2]),
+            BtcECKey.fromPublicOnly(publicKeyBytes[3]),
+            BtcECKey.fromPublicOnly(publicKeyBytes[4]),
+            BtcECKey.fromPublicOnly(publicKeyBytes[5]),
+        });
+        List<FederationMember> members = FederationTestUtils.getFederationMembersWithBtcKeys(keys);
+        PendingFederation pendingFederation = new PendingFederation(members);
+
+        byte[] result = pendingFederation.serialize(activations);
+        StringBuilder expectedBuilder = new StringBuilder();
+        expectedBuilder.append("f8cc");
+        pendingFederation.getBtcPublicKeys().stream().sorted(BtcECKey.PUBKEY_COMPARATOR).forEach(key -> {
+            expectedBuilder.append("a1");
+            expectedBuilder.append(ByteUtil.toHexString(key.getPubKey()));
+        });
+
+        String expected = expectedBuilder.toString();
+        Assertions.assertEquals(expected, ByteUtil.toHexString(result));
+    }
+
+    @Test
+    void deserializePendingFederationOnlyBtcKeys() throws Exception {
+        byte[][] publicKeyBytes = Arrays.asList(100, 200, 300, 400, 500, 600).stream()
+            .map(k -> BtcECKey.fromPrivate(BigInteger.valueOf(k)))
+            .sorted(BtcECKey.PUBKEY_COMPARATOR)
+            .map(k -> k.getPubKey())
+            .toArray(byte[][]::new);
+
+        byte[][] rlpBytes = new byte[publicKeyBytes.length][];
+
+        for (int k = 0; k < publicKeyBytes.length; k++) {
+            rlpBytes[k] = RLP.encodeElement(publicKeyBytes[k]);
+        }
+
+        byte[] data = RLP.encodeList(rlpBytes);
+
+        PendingFederation deserializedPendingFederation = PendingFederation.deserializeOnlyBtcKeys(data);
+
+        Assertions.assertEquals(6, deserializedPendingFederation.getBtcPublicKeys().size());
+        for (int i = 0; i < 6; i++) {
+            Assertions.assertTrue(Arrays.equals(publicKeyBytes[i], deserializedPendingFederation.getBtcPublicKeys().get(i).getPubKey()));
+        }
+    }
+
     private void testBuildFederation(
         boolean isRskip201Active,
         boolean isRskip284Active,
@@ -315,5 +451,9 @@ class PendingFederationTest {
         if (isRskip201Active && !isRskip284Active && networkId.equals(NetworkParameters.ID_TESTNET)) {
             assertEquals(TestConstants.ERP_TESTNET_REDEEM_SCRIPT, builtFederation.getRedeemScript());
         }
+    }
+
+    private int randomInRange(int min, int max) {
+        return TestUtils.generateInt(PendingFederationTest.class.toString(),max - min + 1) + min;
     }
 }
